Remove input types for DerivedQuery
Why?

Well we don't _actually_ know the input type to the derived query as the prior derivation could be optimized down to
returning a `DBResult` -or- it could not be optimized and instead returns a `Model` result.

We could use a union type to express this.. but the type information from the generics all gets lots once
we enter the optimization phase anyway so tracking `TIn` doesn't seem to provide any actual benefits.