start updating code-gen steps to use new schema representation
Shifting to a new format to define schemas.

No longer do we do:

`edgeName: FieldEdge(toType).inverse(ForeignKeyEdge())`

but instead define edges in separate `inboundEdges` and `outboundEdges` sections. This is less confusing than invertin imo.

We also no longer ask the user to declare if it is a field edge or foreign key edge.

They'll just say `Edge` and if the edge is _through_ a field we'll infer the type of edge it is.

Also started moving extra configuration to the `Extensions` object.

There's currently type errors -- but as I'm the only one working on this 🤷‍♂️