change inline edge representation
Former:
```
{
  src: QualifiedNodeReferenceOrType;
  dest: QualifiedNodeReferenceOrType;
}
```

After:
```
{
  throughOrTo: QualifiedNodeReferenceOrType;
}
```

We don't need `src` and `dest` given that `src` is, for outbound edges, always known.
We also always know the `dest` for inbound edges.

The other thing (src or dest) is provided in `throughOrTo`

`through` if we go `through` a field to arrive at the other thing.
`to` if we are just referencing a type.

`to` would be used for backends that do not rely on fields holding edge informaiton.
`through` when the edge information is stored in a field on the src or dest type.

field edge when on src type.
foreign key edge when on dest type.